refactor: create single speed train domain directly
