+++ src/libecalc/domain/process/compressor/core/train/single_speed_compressor_train_common_shaft.py
@@ -2,6 +2,10 @@ from libecalc.common.energy_model_type import EnergyModelType
 from libecalc.common.errors.exceptions import IllegalStateException
 from libecalc.common.fixed_speed_pressure_control import FixedSpeedPressureControl
 from libecalc.common.logger import logger
+from libecalc.domain.component_validation_error import (
+    ProcessChartTypeValidationException,
+    ProcessDischargePressureValidationException,
+)
 from libecalc.domain.process.compressor.core.results import CompressorTrainResultSingleTimeStep
 from libecalc.domain.process.compressor.core.train.base import CompressorTrainModel
 from libecalc.domain.process.compressor.core.train.train_evaluation_input import CompressorTrainEvaluationInput
@@ -12,6 +16,7 @@ from libecalc.domain.process.compressor.core.train.utils.numeric_methods import 
 )
 from libecalc.domain.process.compressor.core.utils import map_compressor_train_stage_to_domain
 from libecalc.domain.process.compressor.dto import CompressorStage
+from libecalc.domain.process.value_objects.chart.compressor import SingleSpeedCompressorChart
 from libecalc.domain.process.value_objects.fluid_stream.fluid_factory import FluidFactoryInterface
 
 
@@ -81,6 +86,8 @@ class SingleSpeedCompressorTrainCommonShaft(CompressorTrainModel):
             maximum_discharge_pressure=maximum_discharge_pressure,
             calculate_max_rate=calculate_max_rate,
         )
+        self._validate_maximum_discharge_pressure()
+        self._validate_stages(stages_mapped)
 
     def evaluate_given_constraints(
         self,
@@ -409,3 +416,17 @@ class SingleSpeedCompressorTrainCommonShaft(CompressorTrainModel):
                 return 0.0
 
         return maximum_mass_rate
+
+    def _validate_maximum_discharge_pressure(self):
+        if self.maximum_discharge_pressure is not None and self.maximum_discharge_pressure < 0:
+            msg = f"maximum_discharge_pressure must be greater than or equal to 0. Invalid value: {self.maximum_discharge_pressure}"
+
+            raise ProcessDischargePressureValidationException(message=str(msg))
+
+    def _validate_stages(self, stages):
+        for stage in stages:
+            if not isinstance(stage.compressor_chart, SingleSpeedCompressorChart):
+                msg = "Single Speed Compressor train only accepts Single Speed Compressor Charts."
+                f" Given type was {type(stage.compressor_chart)}"
+
+                raise ProcessChartTypeValidationException(message=str(msg))
